Refactoring. Adding interfaces
