+++ src/main/java/game/solver/SudokuSolverInterface.java
@@ -1,0 +1,6 @@
+package game.solver;
+
+public interface SudokuSolverInterface {
+
+    boolean solveByBacktracking();
+}
